nevigation activity를 findActivity에 합침
